// src/app/models/adicional.ts
export interface RequestAdicional {
    SDT_SerPreTFST073_Cab: Adicional[];
}

export interface Adicional {
    FS01CONTPO?: string;
    FS46CONNRO?: number;
    FS44SrvIde?: string;
    FS44CodPre?: string;
    FS73NroLin?: string;
    FS73AdiFch?: string;
    FS73AdiCnt?: string;
    FS73MonCnu?: string;
    FS73Imp?: string;
    FS73Tpo?: string;
    FS73DtoImp?: string;
    FS73FchIni?: string;
    FS73FchFin?: string;
    FS73CliCod?: string;
    FS73CliSuc?: string;
    FS73ConIns?: string;
    FS73OrdIns?: string;
    FS73ProRrt?: string;
    FS73InhPrm?: string;
    FS73MonCuo?: string;
    FS73ImpCuo?: string;
    FS73CntCuo?: string;
    FS73FacCuo?: string;
    FS73UsuUpd?: string;
    FS73FhoUpd?: string;
    FS73DMOCNU?: string;
    UserWs?: string;
    PassWs?: string;
    ModoRest?: string;
    Usuario?: string;
}